Rename ParquetChannelConverter to TelemetryChannelSerializer and add IAsyncEnumerable support for converting channels
